CameraV2/requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4

# IMU Bridge dependencies
pyserial>=3.5
websockets>=12.0

# API Server dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0
